app/requirements.txt
# secrets
python-dotenv

# data
pandas

#db
psycopg2-binary
sqlalchemy==1.4.47

#web
flask
flask-sqlalchemy==3.0.5
flask-api